pyproject.toml
[project]
name = "finance-collector"
version = "0.1.0"
description = "Finance information collector API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "yfinance>=0.2.31",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.4.2",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
